requirements.txt
line-profiler==3.5.1
numpy==1.22.3
Pillow==9.1.0
